moved code for getting metric values to a new file
